// src/api/schema.d.ts
export interface components {
    schemas: {
        CatalogCreateRequest: {
            /** @enum {string} */
            type: "ACTION" | "STATUS" | "EVENT_FAILURE" | "PERSONNEL" | "SUPERVISOR";
            value: string;
        };
        CatalogResponse: {
            active?: boolean;
            /** Format: int64 */
            id?: number;
            /** @enum {string} */
            type?: "ACTION" | "STATUS" | "EVENT_FAILURE" | "PERSONNEL" | "SUPERVISOR";
            value?: string;
        };
        CatalogUpdateRequest: {
            active: boolean;
            value: string;
        };
        ClientRequest: {
            /** Format: email */
            contactEmail?: string;
            documentNumber?: string;
            /** @enum {string} */
            documentType?: "RUC" | "DNI";
            name: string;
        };
        ClientResponse: {
            contactEmail?: string;
            documentNumber?: string;
            /** @enum {string} */
            documentType?: "RUC" | "DNI";
            /** Format: int64 */
            id?: number;
            name?: string;
        };
        EquipmentRequest: {
            model: string;
            serialNumber?: string;
        };
        EquipmentResponse: {
            /** Format: int64 */
            clientId?: number;
            clientName?: string;
            /** Format: int64 */
            id?: number;
            model?: string;
            serialNumber?: string;
        };
        LoginRequest: {
            /** Format: email */
            email: string;
            password: string;
        };
        LoginResponse: {
            accessToken?: string;
            /** Format: int64 */
            expiresIn?: number;
            tokenType?: string;
        };
        TechnicalReportRequest: {
            /** Format: int64 */
            actionId: number;
            affectedComponent?: string;
            details: string;
            /** Format: date-time */
            endDatetime: string;
            /** Format: int64 */
            equipmentId: number;
            /** Format: int64 */
            eventFailureId: number;
            /** Format: int64 */
            finalStatusId: number;
            /** Format: int64 */
            initialStatusId: number;
            /** Format: int64 */
            personnelId: number;
            /** Format: date-time */
            startDatetime?: string;
            /** Format: int64 */
            supervisorId: number;
        };
        TechnicalReportResponse: {
            action?: components["schemas"]["CatalogResponse"];
            affectedComponent?: string;
            /** Format: int64 */
            clientId?: number;
            clientName?: string;
            /** Format: date-time */
            createdAt?: string;
            createdBy?: string;
            details?: string;
            /** Format: date-time */
            endDatetime?: string;
            /** Format: int64 */
            equipmentId?: number;
            equipmentModel?: string;
            equipmentSerialNumber?: string;
            eventFailure?: components["schemas"]["CatalogResponse"];
            finalStatus?: components["schemas"]["CatalogResponse"];
            /** Format: int64 */
            id?: number;
            initialStatus?: components["schemas"]["CatalogResponse"];
            personnel?: components["schemas"]["CatalogResponse"];
            reportNumber?: string;
            /** Format: date-time */
            startDatetime?: string;
            supervisor?: components["schemas"]["CatalogResponse"];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}